fix(user-repo): rename repository methods to match actual entity property names (userName, userEmail)
