+++ src/main/java/com/f5/buzon_inteligente_BE/user/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface UserRepository extends JpaRepository<User, Long>{
 
     Optional<User> findById(Long id);
-    Optional<User> findByEmail(String email);
-    Optional<User> findByUsername(String username);
+    Optional<User> findByUserEmail(String userEmail);
+    Optional<User> findByUserName(String userName);
     Optional<User> findByPermanentCredential(String permanentCredential);
 }
