+++ src/main/resources/application.properties
@@ -4,17 +4,18 @@ spring.application.name=buzon-inteligente-BE
 
 
 
-spring.datasource.url=jdbc:mysql://localhost:3306/buzoninteligente?createDatabaseIfNotExist=true&useSSL=false&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/buzoninteligente
 spring.datasource.username=${DATABASE_USERNAME}
 spring.datasource.password=${DATABASE_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.defer-datasource-initialization=true
-spring.jpa.show-sql: true
+spring.jpa.show-sql=true
 
 spring.sql.init.mode=always
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.database-platform=org.hibernate.dialect.MySQLDialect
 
-jwt.secret=LTYf9QA/oWeanNqjAt1j+1xikh08dUKEqHp/Ca0rQpg=
-jwt.expiration=86400000
+
+jwt.secret=${JWT_SECRET}
+jwt.expiration=${JWT_EXPIRATION}
